scripts/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
opencv-python
torch
torchvision
langchain
langchain-core>=0.1.1
langchain-community
langchain-groq
langchain-openai
langchain-text-splitters
groq
pydantic
pytube
pillow
numpy
markdown
openai
python-dotenv
requests
aiohttp
faiss-cpu
matplotlib
python-jose 